--- FileHelpers/Engines/MultiRecordEngine.docs.xml
@@ -1,158 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?> 
-<doc>
-	<ReadFile>
-		<summary>
-		Read a file and return an array of the contained records.
-		</summary>
-		<remarks>
-			This method opens, reads and closes the file
-			(don't open or close the file before or after calling this method)
-		</remarks>
-		<include file='Examples.xml' path='doc/examples/ReadFile/*'/>
-		<param name="fileName">The file path to be read.</param>
-		<returns>An array of the records in the file</returns>
-	</ReadFile>
-
-	<ReadStream>
-		<summary>
-		Read a Stream and return an array of the contained records.
-		</summary>
-		<remarks>
-		This method only uses the stream and does not close the stream after using it, you must do it.
-		</remarks>
-		<include file='Examples.xml' path='doc/examples/ReadFile/*'/>
-		<param name="reader">The reader of the source stream.</param>
-		<returns>An array of the records in the Stream</returns>
-	</ReadStream>
-
-	<ReadString>
-		<summary>
-		Read a String and return an array of the contained records.
-		</summary>
-		<param name="source">The string that contains the records.</param>
-		<returns>An array of the records in the String.</returns>
-	</ReadString>
-
-	<AppendToFile1>
-		<summary>
-		Append a record to the specified file.
-		</summary>
-		<remarks>
-			This method opens, seeks to the end, writes and closes the file
-			(don't open or close the file before or after calling this method)
-		</remarks>
-		<include file='Examples.xml' path='doc/examples/AppendFile/*'/>
-		<param name="fileName">The file path to be written at end.</param>
-		<param name="record">The record to write.</param>
-		<returns>True if the operation is successful. False otherwise.</returns>
-	</AppendToFile1>
-
-	<AppendToFile2>
-		<summary>
-		Append an array of records to the specified file.
-		</summary>
-		<remarks>
-			This method opens, seeks to the end, writes and closes the file
-			(don't open or close the file before or after calling this method)
-		</remarks>
-		<include file='Examples.xml' path='doc/examples/AppendFile/*'/>
-	
-		<param name="fileName">The file path to be written at end.</param>
-		<param name="records">The records to write (Can be an array, ArrayList, etc)</param>
-		<returns>True if the operation is successful. False otherwise.</returns>
-	</AppendToFile2>
-
-	<WriteFile>
-		<summary>
-		Write an array of records to the specified file.
-		</summary>
-		<remarks>
-			<para>This method opens, writes and closes the file
-			(don't open or close the file before or after calling this method)</para>
-		<para>This method over writes existing files.</para>
-		</remarks>
-		<include file='Examples.xml' path='doc/examples/WriteFile/*'/>
-		<param name="fileName">The file path to be write.</param>
-		<param name="records">The records to write (Can be an array, ArrayList, etc)</param>
-		<returns>True if the operation is successful. False otherwise.</returns>
-	</WriteFile>
-
-	<WriteFile2>
-		<summary>
-		Write the specified number of records from the array to a file.
-		</summary>
-		<remarks>
-			<para>This method opens, writes and closes the file
-			(don't open or close the file before or after calling this method)</para>
-		<para>This method over writes existing files.</para>
-		</remarks>
-		<include file='Examples.xml' path='doc/examples/WriteFile/*'/>
-		<param name="fileName">The file path to be write.</param>
-		<param name="records">The records to write (Can be an array, ArrayList, etc)</param>
-		<param name="maxRecords">The max number of array elements to write.</param>
-		<returns>True if the operation is successful. False otherwise.</returns>
-	</WriteFile2>
-
-	<WriteStream>
-		<summary>
-		Write an array of records to the specified Stream.
-		</summary>
-		<remarks>
-		This method only uses the stream and does not close the stream after using it, you must do it.
-		</remarks>
-		<include file='Examples.xml' path='doc/examples/WriteFile/*'/>
-		<param name="writer">The writer of the source stream.</param>
-		<param name="records">The records to write (Can be an array, ArrayList, etc)</param>
-		<returns>True if the operation is successful. False otherwise.</returns>
-	</WriteStream>
-
-	<WriteStream2>
-		<summary>
-		Write the specified number of records in the array to the Stream.
-		</summary>
-		<remarks>
-		This method only uses the stream and does not close the stream after using it, you must do it.
-		</remarks>
-		<include file='Examples.xml' path='doc/examples/WriteFile/*'/>
-		<param name="writer">The writer of the source stream.</param>
-		<param name="records">The records to write (Can be an array, ArrayList, etc)</param>
-		<param name="maxRecords">The max number of array elements to write.</param>
-		<returns>True if the operation is successful. False otherwise.</returns>
-	</WriteStream2>
-
-	<WriteString>
-		<summary>
-		Write an array of records to an String and return it.
-		</summary>
-		<param name="records">The records to write (Can be an array, ArrayList, etc)</param>
-		<returns>The resulting string after write the records.</returns>
-	</WriteString>
-
-	<WriteString2>
-		<summary>
-		Write an array of records to an String and return it.
-		</summary>
-		<param name="records">The records to write (Can be an array, ArrayList, etc)</param>
-		<returns>The resulting string after write the records.</returns>
-		<param name="maxRecords">The max number of array elements to write.</param>
-		<returns>True if the operation is successful. False otherwise.</returns>
-	</WriteString2>
-
-	<LineNum>
-		<summary>The current line number.</summary>
-	</LineNum>
-
-	<ColumnNum>
-		<summary>The current column number.</summary>
-	</ColumnNum>
-
-	<TotalRecords>
-		<summary>The total numbers of records in the last file read or written
-			(only works with whole read/write).</summary>
-	</TotalRecords>
-
-	<RecordType>
-		<summary>Returns the type of records handled by this engine.</summary>
-	</RecordType>
-
-</doc>
